requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
gurobipy==9.5.0
idna==3.3
numpy==1.22.1
pandas==1.4.1
pycodestyle==2.8.0
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
XlsxWriter==3.0.2
